work

+++ src/WebIDL/alpha_wrap_glue.cpp
@@ -27,19 +27,19 @@ AlphaWrap* EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_AlphaWrap_0() {
   return new AlphaWrap();
 }
 
-AlphaWrap* EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_AlphaWrap_2(double alpha, double offset) {
+AlphaWrap* EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_AlphaWrap_2(float alpha, float offset) {
   return new AlphaWrap(alpha, offset);
 }
 
-void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_setAlpha_1(AlphaWrap* self, double alpha) {
+void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_setAlpha_1(AlphaWrap* self, float alpha) {
   self->setAlpha(alpha);
 }
 
-void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_setOffset_1(AlphaWrap* self, double offset) {
+void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_setOffset_1(AlphaWrap* self, float offset) {
   self->setOffset(offset);
 }
 
-void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_addPoint_3(AlphaWrap* self, double x, double y, double z) {
+void EMSCRIPTEN_KEEPALIVE emscripten_bind_AlphaWrap_addPoint_3(AlphaWrap* self, float x, float y, float z) {
   self->addPoint(x, y, z);
 }
